# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TNC Trading API
  description: |
    API de la plateforme TNC Trading - Tokenisation d'Or Souveraine pour le Burkina Faso.
    
    ## Authentification
    L'API utilise des JWT Bearer tokens. Après login, incluez le token dans le header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requêtes/minute par IP
    - 1000 requêtes/minute par utilisateur authentifié
    
    ## Codes d'erreur
    Toutes les erreurs suivent le format standard avec un code et message.
  version: 1.0.0
  contact:
    name: TNC Trading Support
    email: support@tnc-trading.com
  license:
    name: Proprietary
servers:
  - url: https://api.tnc-trading.com/api/v1
    description: Production
  - url: https://staging-api.tnc-trading.com/api/v1
    description: Staging
  - url: http://localhost:8787/api/v1
    description: Development
...
paths:
...
  /market/price/history:
    get:
      tags: [Market]
      summary: Historique des prix
      operationId: getPriceHistory
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d, 1y]
            default: 7d
        - name: interval
          in: query
          schema:
            type: string
            enum: [1h, 4h, 1d]
            default: 1h
      responses:
        '200':
          description: Historique des prix
          content:
            application/json:
              schema:
                type: object
                properties:
                  prices:
                    type: array
                    items:
                      $ref: '#/components/schemas/GoldPriceHistory'
...
  /market/quote:
    post:
      tags: [Market]
      summary: Obtenir un devis pour achat/vente
      operationId: getQuote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Devis valide 60 secondes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
  /market/buy:
    post:
      tags: [Market]
      summary: Acheter des tokens d'or
      operationId: buyTokens
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyRequest'
      responses:
        '200':
          description: Redirection vers le paiement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRedirect'
        '400':
          description: Stock insuffisant ou limite dépassée
  /market/sell:
    post:
      tags: [Market]
      summary: Vendre des tokens d'or
      operationId: sellTokens
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellRequest'
      responses:
        '200':
          description: Vente effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

  # ==================== WALLET ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cloudflareAccess:
      type: apiKey
      in: header
      name: CF-Access-JWT-Assertion
      description: Cloudflare Access JWT
...
  schemas:
...
    GoldPriceHistory:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        priceXof:
          type: number
        priceUsd:
          type: number
...
    QuoteRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [BUY, SELL]
        tokenAmount:
          type: number
          minimum: 1
          description: Quantité en grammes (obligatoire si pas de cashAmount)
        cashAmount:
          type: number
          minimum: 0
          description: Montant en XOF (obligatoire si pas de tokenAmount)
    Quote:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [BUY, SELL]
        tokenAmount:
          type: number
        cashAmount:
          type: number
        pricePerGram:
          type: number
        fees:
          type: number
        total:
          type: number
        validUntil:
          type: string
          format: date-time
    BuyRequest:
      type: object
      required: [quoteId, paymentMethod]
      properties:
        quoteId:
          type: string
        paymentMethod:
          type: string
          enum: [orange_money, moov_money, card, cinetpay]
        phoneNumber:
          type: string
          description: Pour Mobile Money
    SellRequest:
      type: object
      required: [quoteId, twoFactorCode]
      properties:
        quoteId:
          type: string
        twoFactorCode:
          type: string
          pattern: '^\d{6}$'
    PaymentRedirect:
      type: object
      properties:
        transactionId:
          type: string
        paymentUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time

    # ===== Wallet =====
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [BUY, SELL, DEPOSIT, WITHDRAWAL, FEE]
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
        tokenAmount:
          type: number
        cashAmount:
          type: number
        pricePerGram:
          type: number
        fees:
          type: number
        paymentMethod:
          type: string
        paymentReference:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...